+++ features/preferences/impl/src/main/kotlin/io/element/android/features/preferences/impl/advanced/AdvancedSettingsEvents.kt
@@ -19,7 +19,6 @@ package io.element.android.features.preferences.impl.advanced
 import io.element.android.compound.theme.Theme
 
 sealed interface AdvancedSettingsEvents {
-    data class SetRichTextEditorEnabled(val enabled: Boolean) : AdvancedSettingsEvents
     data class SetDeveloperModeEnabled(val enabled: Boolean) : AdvancedSettingsEvents
     data class SetSharePresenceEnabled(val enabled: Boolean) : AdvancedSettingsEvents
     data object ChangeTheme : AdvancedSettingsEvents
