+++ libraries/preferences/test/src/main/kotlin/io/element/android/libraries/preferences/test/InMemoryAppPreferencesStore.kt
@@ -21,23 +21,13 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 
 class InMemoryAppPreferencesStore(
-    isRichTextEditorEnabled: Boolean = false,
     isDeveloperModeEnabled: Boolean = false,
     customElementCallBaseUrl: String? = null,
     theme: String? = null,
 ) : AppPreferencesStore {
-    private val isRichTextEditorEnabled = MutableStateFlow(isRichTextEditorEnabled)
     private val isDeveloperModeEnabled = MutableStateFlow(isDeveloperModeEnabled)
     private val customElementCallBaseUrl = MutableStateFlow(customElementCallBaseUrl)
     private val theme = MutableStateFlow(theme)
-
-    override suspend fun setRichTextEditorEnabled(enabled: Boolean) {
-        isRichTextEditorEnabled.value = enabled
-    }
-
-    override fun isRichTextEditorEnabledFlow(): Flow<Boolean> {
-        return isRichTextEditorEnabled
-    }
 
     override suspend fun setDeveloperModeEnabled(enabled: Boolean) {
         isDeveloperModeEnabled.value = enabled
